+++ CHANGELOG.md
@@ -12,6 +12,7 @@ when they are specified in the stubbing, but not in the mentioned `throws` claus
 - [#22](https://github.com/picimako/mockitools/issues/22): Added new action to convert from and to `InOrder.verify()` based verification.
 Along with it extended the possible conversion options from and to `Mockito.verify()` and `BDDMockito.then()` as well.
 Details are available in the [Verifications](https://github.com/picimako/mockitools/blob/main/docs/verifications.md) document.
+- `EnforceConventionInspection` can now report `InOrder.verify()` calls too, as part of enforcing `BDDMockito.then()` verification.
 
 ### Changed
 - Removed support for IJ2021.1. From now on 2021.2 is the earliest version supported.